finished refining youtube_get_all_comments_and_parse.py
